+++ Changelog.md
@@ -2,6 +2,32 @@
 
 Changelog for JwtAuthenticationApi System. 
 
+## [0.3.0] - 07.05.2023 - JAA-3-ImplementLoggingUsingSerilog
+
+### Added
+- NuGet packages: Serilog, Polly, Serilog.AspNetCore, Serilog.Enrichers.Thread, Serilog.Sinks.Async, Serilog.Sinks.Console, 
+Serilog.Sinks.File, SerilogWeb.Classic
+- LoggerSetup class that setup ILogger from Serilog.
+- Logging in SaltService.
+- Retry policy in SaltService using Polly.
+- Tests for all new classes.
+- New tests for polly in SaltService.
+- PollySleepingIntervalsFactory class that creates sleeping intervals for Polly.
+- Semaphore wrapper and semaphore wrapper factory.
+- RegisterServices method in ContainerSetup.
+- SettingsNotProvided Exception.
+- ReturnHttpNotAcceptable option in setup.
+- Try-Catch statements in SaltService. 
+
+### Changed
+- Lock in SaltService is now done using Semaphore.
+- All tests received "Pararelizable" attribute.
+- Tests in SaltService.
+- Switched to Widnows VM in github actions to gain ability to use semaphore.
+
+### Removed
+- Microsoft.Extensions.Logging.ILogger is no longer basic logging system.
+
 ## [0.2.0] - 05.03.2023 - JAA-2-ImplementPasswordHashing
 
 ### Added
@@ -23,7 +49,7 @@ Changelog for JwtAuthenticationApi System.
 ### Changed
 - Container is now set up using extensions methods.
 - Updated Readme.md.
-~~~~
+
 ### Removed
 - WeatherForecastController with its endpoints.